Refactor to move code to logical places
