+++ src/db/crud.py
@@ -1,5 +1,5 @@
 import random
-import helpers
+import src.helpers as helpers
 
 from typing import List
 from sqlalchemy import delete, and_, or_
