+++ src/helpers.py
@@ -1,5 +1,4 @@
-from db import models, database
-from slack_sdk import WebClient
+from src.db import models
 from typing import List, Tuple
 
 
@@ -12,14 +11,6 @@ def generate_member_model_list(member_ids, exclude_members, channel_id, team_id)
             )
 
     return members
-
-
-def get_slack_client(enterprise_id: str | None, team_id: str) -> WebClient:
-    installation = database.installation_store.find_installation(
-        enterprise_id=enterprise_id, team_id=team_id
-    )
-
-    return WebClient(token=installation.bot_token)
 
 
 def round_robin_match(members: List) -> Tuple[List[List], List]:
